# docs/swagger.yaml
basePath: /
definitions:
...
  ttlock.LockAutoLockTimeRequestParams:
    properties:
      accessToken:
        type: string
      clientId:
        type: string
      date:
        type: integer
      lockId:
        type: integer
      seconds:
        type: integer
      type:
        enum:
        - 1
        - 2
        type: integer
    required:
    - accessToken
    - clientId
    - date
    - lockId
    - seconds
    - type
    type: object
  ttlock.LockChangeAdminPasscodeRequestParams:
    properties:
      accessToken:
        type: string
      changeType:
        enum:
        - 1
        - 2
        type: integer
      clientId:
        type: string
      date:
        type: integer
      lockId:
        type: integer
      password:
        type: string
    required:
    - accessToken
    - changeType
    - clientId
    - date
    - lockId
    - password
    type: object
  ttlock.LockDeleteRequestParams:
    properties:
      accessToken:
        type: string
      clientId:
        type: string
      date:
        type: integer
      lockId:
        type: integer
    required:
    - accessToken
    - clientId
    - date
    - lockId
    type: object
  ttlock.LockDetailsRequestParams:
    properties:
      accessToken:
        type: string
      clientId:
        type: string
      date:
        type: integer
      lockId:
        type: integer
    required:
    - accessToken
    - clientId
    - date
    - lockId
    type: object
  ttlock.LockInitializeRequest:
    properties:
      accessToken:
        type: string
      clientId:
        type: string
      date:
        type: integer
      groupId:
        type: integer
      lockAlias:
        type: string
      lockData:
        type: string
      nbInitSuccess:
        type: integer
    required:
    - accessToken
    - clientId
    - date
    - lockData
    type: object
  ttlock.LockListEKeyRequestParams:
    properties:
      accessToken:
        type: string
      clientId:
        type: string
      date:
        type: integer
      keyRight:
        enum:
        - 1
        - 2
        type: integer
      lockId:
        type: integer
      orderBy:
        enum:
        - 1
        - 2
        type: integer
      pageNo:
        minimum: 1
        type: integer
      pageSize:
        maximum: 200
        minimum: 1
        type: integer
      searchStr:
        type: string
    required:
    - accessToken
    - clientId
    - date
    - lockId
    - pageNo
    - pageSize
    type: object
  ttlock.LockListRequestParams:
    properties:
      accessToken:
        type: string
      clientId:
        type: string
      date:
        type: integer
      groupId:
        type: integer
      lockAlias:
        type: string
      pageNo:
        minimum: 1
        type: integer
      pageSize:
        maximum: 1000
        minimum: 1
        type: integer
    required:
    - accessToken
    - clientId
    - date
    - pageNo
    - pageSize
    type: object
  ttlock.LockRenameRequestParams:
    properties:
      accessToken:
        type: string
      clientId:
        type: string
      date:
        type: integer
      lockAlias:
        type: string
      lockId:
        type: integer
    required:
    - accessToken
    - clientId
    - date
    - lockAlias
    - lockId
    type: object
  ttlock.LockUpdateRequestParams:
    properties:
      accessToken:
        type: string
      clientId:
        type: string
      date:
        type: integer
      lockData:
        type: string
      lockId:
        type: integer
    required:
    - accessToken
    - clientId
    - date
    - lockData
    - lockId
    type: object
...
  ttlock.PasscodeListRequestParams:
    properties:
      accessToken:
        type: string
      clientId:
        type: string
      date:
        type: integer
      lockId:
        type: integer
      orderBy:
        enum:
        - 1
        - 2
        type: integer
      pageNo:
        minimum: 1
        type: integer
      pageSize:
        maximum: 200
        minimum: 1
        type: integer
      searchStr:
        type: string
    required:
    - accessToken
    - clientId
    - date
    - lockId
    - orderBy
    - pageNo
    - pageSize
    type: object
...
info:
  contact: {}
  title: TTLock Api
  version: "1.0"
paths:
...
  /api/locks/auto-lock-time:
    post:
      consumes:
      - application/json
      description: Update TTLock auto lock time
      parameters:
      - description: Lock auto lock time request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.LockAutoLockTimeRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: Lock auto lock time updated successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Set lock auto lock time
      tags:
      - Locks
  /api/locks/change-admin-passcode:
    post:
      consumes:
      - application/json
      description: Change TTLock admin passcode
      parameters:
      - description: Lock admin passcode request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.LockChangeAdminPasscodeRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: Lock admin passcode changed successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Change lock admin passcode
      tags:
      - Locks
  /api/locks/delete:
    post:
      consumes:
      - application/json
      description: Delete a TTLock
      parameters:
      - description: Lock delete request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.LockDeleteRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: Lock deleted successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Delete lock
      tags:
      - Locks
  /api/locks/detail:
    post:
      consumes:
      - application/json
      description: Get TTLock details
      parameters:
      - description: Lock detail request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.LockDetailsRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: Lock detail fetched successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Get lock detail
      tags:
      - Locks
  /api/locks/initialize:
    post:
      consumes:
      - application/json
      description: Initialize a TTLock device and register it in the system
      parameters:
      - description: Lock initialize payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.LockInitializeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Lock initialized successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Initialize a TTLock
      tags:
      - Locks
  /api/locks/keys:
    post:
      consumes:
      - application/json
      description: Get all eKeys associated with a TTLock
      parameters:
      - description: Lock eKeys list request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.LockListEKeyRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: Lock eKeys fetched successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Get lock eKeys list
      tags:
      - Locks
  /api/locks/list:
    post:
      consumes:
      - application/json
      description: Get a list of TTLocks
      parameters:
      - description: Lock list request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.LockListRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: Lock list fetched successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Get lock list
      tags:
      - Locks
  /api/locks/passcodes:
    post:
      consumes:
      - application/json
      description: Get all passcodes associated with a TTLock
      parameters:
      - description: Lock passcodes list request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.PasscodeListRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: Lock passcodes fetched successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Get lock passcodes list
      tags:
      - Locks
  /api/locks/rename:
    post:
      consumes:
      - application/json
      description: Rename a TTLock
      parameters:
      - description: Lock rename request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.LockRenameRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: Lock renamed successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Rename lock
      tags:
      - Locks
  /api/locks/update:
    post:
      consumes:
      - application/json
      description: Update TTLock information
      parameters:
      - description: Lock update request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.LockUpdateRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: Lock updated successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Update lock
      tags:
      - Locks
...
swagger: "2.0"